Update anydesk.yaml
Updated the connection_trace.txt information as it incorrectly stated that the remote IP address would be shown in this log. Only the remote ID is shown in this file. Also updated it to show the additional two entries that may be visible, REJECTED and Token. Also, removed the entries regarding the connection_trace.txt being in the %APPDAATA% folder as it is only in the %PROGRAMDATA% directory.